requirements.txt
pandas
praw
requests
sqlalchemy
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-http
apache-airflow-providers-amazon
apache-airflow-providers-redis
apache-airflow-providers-ssh
apache-airflow-providers-docker
apache-airflow-providers-sqlite
apache-airflow-providers-ftp
s3fs